package: upgrade all dependencies
Upgrade all 'dependencies' and 'devDependencies' packages to their latest
versions, as well as run 'npm audit fix'. This should resolve the issues in
all open Dependabot pull requests, as well as use the latest 'actions/core'
package to eliminate the 'set-output' deprecation warning.

Note that this commit does *not* build successfully. Due to the complexity
of piecewise updates of dependencies, we will upgrade all packages here, and
in subsequent commits update the rest of the repository to handle any
breaking changes associated with those upgrades.

Signed-off-by: Victoria Dye <vdye@github.com>